backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.0
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
emoji==2.15.0
et_xmlfile==2.0.0
fastapi==0.121.1
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.12.0
google-api-core==2.28.1
google-auth==2.41.1
google-cloud-vision==3.11.0
googleapis-common-protos==1.70.0
grandalf==0.8
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
grpcio-status==1.76.0
html5lib==1.1
httptools==0.7.1
httpx-sse==0.4.2
humanfriendly==10.0
idna==3.10
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jiter==0.11.0
joblib==1.5.2
jq==1.10.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.0
langchain-chroma==0.2.6
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-deepseek==1.0.1
langchain-huggingface==1.0.1
langchain-mcp-adapters==0.1.12
langchain-ollama==1.0.0
langchain-openai==1.1.0
langchain-tavily==0.2.12
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langgraph-cli==0.4.3
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.33
lazy_loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
numexpr==2.13.1
oauthlib==3.3.1
olefile==0.47
ollama==0.6.0
omegaconf==2.3.0
openai==1.109.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.12.0
pi_heif==1.1.1
pikepdf==10.0.0
pillow==12.0.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg-pool==3.2.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
PyJWT==2.10.1
PyMySQL==1.1.2
pypandoc==1.15
pyparsing==3.2.5
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.7
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
regex==2025.9.18
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
sse-starlette==3.0.3
starlette==0.49.3
striprtf==0.0.26
sympy==1.14.0
tavily==1.1.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
tqdm==4.67.1
typer==0.19.2
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.37.0
watchfiles==1.1.0
loguru==0.7.2
PyYAML==6.0.2
python-dotenv==1.1.1
pydantic-settings==2.11.0
httpx==0.28.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
passlib==1.7.4
